+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to `sensor-qr-app` will be documented in this file.
 
 ## [Unreleased] -
 ### Added
+- Sensor Register
 - QR Read
 - QR Qenreate
 - GUI